Cleanup
git-svn-id: https://subtitleedit.googlecode.com/svn/trunk@493 99eadd0c-20b8-1223-b5c4-2a2b2df33de2

+++ src/Controls/AudioVisualizer.cs
@@ -407,7 +407,7 @@ namespace Nikse.SubtitleEdit.Controls
                         if (currentRegionWidth > 40)
                         {
                             SolidBrush textBrush = new SolidBrush(Color.Turquoise);
-                            e.Graphics.DrawString(string.Format("{0:0.###} {1}", (double)((double)currentRegionWidth / _wavePeaks.Header.SampleRate / _zoomFactor),
+                            e.Graphics.DrawString(string.Format("{0:0.###} {1}",((double)currentRegionWidth / _wavePeaks.Header.SampleRate / _zoomFactor),
                                                                 Configuration.Settings.Language.WaveForm.Seconds),
                                                   Font, textBrush, new PointF(currentRegionLeft + 3, Height - 32));
                         }
